requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
pillow
pinecone
python-multipart
sqlalchemy
